# Conflicts:
#	src/main/java/com/example/spring_project/controller/MatchController.java

+++ src/main/java/com/example/spring_project/domain/match/MatchRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MatchRepository extends JpaRepository<Match, Integer> {
-    public Match getMatchByNo(int no);
+public interface MatchRepository extends JpaRepository<Match, Long> {
+    public Match getMatchByNo(long no);
     public List<Match> findAllByTitleLike(String pattern, Pageable pageable);
     public List<Match> findAllByStartAtLike(String pattern);
